requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
coinbase-advanced-py==1.0.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0
flask==2.3.3
requests==2.31.0
asyncio>=3.4.3
ta==0.10.2
# Adding required technical analysis libraries
scipy>=1.10.0  # For statistical calculations
statsmodels>=0.14.0  # For time series analysis 